+++ Discord_bot.py
@@ -1,3 +1,4 @@
+from asyncio import run
 from sys import platform, exit, argv
 from threading import enumerate as threads
 from traceback import format_exc
@@ -72,8 +73,13 @@ def main():
             Config.read_config(change_servers=(len(argv) > 1 and argv[1] == "-cs"))
             setup_print_handlers()
         bot = commands.Bot(command_prefix=get_prefix, intents=Intents.all(), help_command=None)
-        for i in [Poll, ChatCommands, MinecraftCommands]:
-            bot.add_cog(i(bot))
+
+        async def add_cogs(bot: commands.Bot):
+            for i in [Poll, ChatCommands, MinecraftCommands]:
+                await bot.add_cog(i(bot))
+            return bot
+
+        bot = run(add_cogs(bot))
         create_pot_lines(bot)
         print(get_translation("Bot started!"))
         BotVars.bot_for_webhooks = bot